hotel currency code and debuging hottel policy

+++ service/src/main/java/com/eligasht/service/model/newModel/flight/confirmFlightPrice/response/BaseFare.java
@@ -11,7 +11,7 @@ public class BaseFare {
     private String currencyCode;
     @SerializedName("Amount")
     @Expose
-    private Integer amount;
+    private Double amount;
 
     public String getCurrencyCode() {
         return currencyCode;
@@ -21,11 +21,11 @@ public class BaseFare {
         this.currencyCode = currencyCode;
     }
 
-    public Integer getAmount() {
+    public Double getAmount() {
         return amount;
     }
 
-    public void setAmount(Integer amount) {
+    public void setAmount(Double amount) {
         this.amount = amount;
     }
