+++ service/src/main/java/com/eligasht/service/model/newModel/flight/confirmFlightPrice/response/TotalFareCost.java
@@ -11,7 +11,7 @@ public class TotalFareCost {
     private String currencyCode;
     @SerializedName("Amount")
     @Expose
-    private Integer amount;
+    private Float amount;
 
     public String getCurrencyCode() {
         return currencyCode;
@@ -21,11 +21,11 @@ public class TotalFareCost {
         this.currencyCode = currencyCode;
     }
 
-    public Integer getAmount() {
+    public Float getAmount() {
         return amount;
     }
 
-    public void setAmount(Integer amount) {
+    public void setAmount(Float amount) {
         this.amount = amount;
     }
