+++ service/src/main/java/com/eligasht/service/model/newModel/hotel/getRoom/response/Room.java
@@ -30,13 +30,13 @@ public class Room {
     private List<Object> childAges = null;
     @SerializedName("SourcePrice")
     @Expose
-    private Integer sourcePrice;
+    private Double sourcePrice;
     @SerializedName("OldPrice")
     @Expose
-    private Integer oldPrice;
+    private Double oldPrice;
     @SerializedName("Price")
     @Expose
-    private Integer price;
+    private Double price;
     @SerializedName("CurrencyPrice")
     @Expose
     private Double currencyPrice;
@@ -106,27 +106,27 @@ public class Room {
         this.childAges = childAges;
     }
 
-    public Integer getSourcePrice() {
+    public Double getSourcePrice() {
         return sourcePrice;
     }
 
-    public void setSourcePrice(Integer sourcePrice) {
+    public void setSourcePrice(Double sourcePrice) {
         this.sourcePrice = sourcePrice;
     }
 
-    public Integer getOldPrice() {
+    public Double getOldPrice() {
         return oldPrice;
     }
 
-    public void setOldPrice(Integer oldPrice) {
+    public void setOldPrice(Double oldPrice) {
         this.oldPrice = oldPrice;
     }
 
-    public Integer getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(Integer price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
